Organizing and structuring the code into modules or smaller functions
