+++ utils.js
@@ -1,15 +1,15 @@
-import jwt from 'jsonwebtoken'
-export const generateLogToken = (user) => {
-    return jwt.sign({
-        _id: user.id,
-        fullname: user.fullname,
-        email: user.email
-    },
-        process.env.JWT_PASS || 'mysecretpass',
-        {
-            expiresIn: '1d',
-        }
-    );
-};
+// import jwt from 'jsonwebtoken'
+// export const generateLogToken = (user) => {
+//     return jwt.sign({
+//         _id: user.id,
+//         fullname: user.fullname,
+//         email: user.email
+//     },
+//         process.env.JWT_PASS || 'mysecretpass',
+//         {
+//             expiresIn: '1d',
+//         }
+//     );
+// };
